Resolved code review response

+++ source/blender/draw/engines/eevee_next/shaders/eevee_sampling_lib.glsl
@@ -107,7 +107,7 @@ vec3 sample_sphere(vec2 rand)
   float omega = rand.y * 2.0 * M_PI;
   float cos_theta = rand.x * 2.0 - 1.0;
   float sin_theta = safe_sqrt(1.0 - cos_theta * cos_theta);
-  return vec3(sin_theta * vec2(cosf(omega), sinf(omega)), cos_theta);
+  return vec3(sin_theta * vec2(cos(omega), sin(omega)), cos_theta);
 }
 
 /** \} */
